requirements.txt
h5py==2.10.0
numpy==1.19.5
tensorflow-gpu==2.4.1
python-chess==0.31.4